pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dip_validator"
version = "0.1.0"
description = "High-precision Dip Validator CLI"
requires-python = ">=3.10"
dependencies = [
    "rtmlib>=0.0.14",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
